# openapi.yaml
openapi: 3.1.0
info:
  title: Spigot API
  version: '1.0.0'
  description: |
    Pay-per-call API marketplace settled in USDC on Arc Testnet.

    **Two ways to pay.** Either works against the same endpoints.

    1. **Credits** — sign in, buy credit once from your wallet, then call any listed
       API with a single authenticated request. The gateway signs and settles each
       payment for you, so there is no wallet popup per call and you never hold gas tokens.
    2. **Direct Nanopayments** — call endpoints with signed nanopayment authorization.
       The gateway / facilitator covers gas, so the caller needs no native gas tokens either.

    **Generating a client.** This document is the source of truth. Any OpenAPI
    generator will produce a client in your language:

    ```bash
    npx @openapitools/openapi-generator-cli generate \
      -i https://raw.githubusercontent.com/0xSarnavo/spigot-arc/main/openapi.yaml \
      -g python -o ./arc-client
    ```
  license:
    name: Proprietary — all rights reserved
  contact:
    name: Spigot
    url: https://github.com/0xSarnavo/spigot-arc
servers:
  - url: http://localhost:4402
    description: Local development
  - url: https://your-deployment.example.com
    description: Your deployment
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/apis/{id}/publish:
    put:
      tags: [Publishing]
      summary: Go live
      description: |
        Requires endpoints, pricing and a payout address. The upstream is probed first and
        rejected if it demands its own payment, because the caller would then be charged
        twice and receive a payment demand instead of data.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Published }
        '400': { description: Incomplete, or the upstream is already x402-protected }

  # ---------------------------------------------------------- Publishers ----
  /api/publishers:
    post:
      tags: [Publishing]
      summary: Become a publisher
      description: Requires a session. Your email is taken from it rather than the request body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                company: { type: string }
                website: { type: string, format: uri }
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  api_key:
                    type: string
                    description: For server-to-server clients. Browsers should use the session.
  /api/publishers/{id}/wallet:
    post:
      tags: [Publishing]
      summary: Set or change your payout address
      description: |
        Setting it the first time is immediate. **Changing** it requires a confirmation code
        emailed to you, so a stolen credential cannot silently reroute your revenue: call
        once without `confirmationCode` to trigger the email, then again with it.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address: { type: string, description: Arc address }
                min_payout: { type: number, default: 10 }
                confirmationCode: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200': { description: Saved }
        '202': { description: Confirmation emailed; resend with confirmationCode }
        '400': { description: Not a valid Arc address }

  # ----------------------------------------------------------- Dashboard ----
  /api/dashboard/{publisherId}:
    get:
      tags: [Dashboard]
      summary: Traffic and revenue
      description: Pass `active` to use the publisher tied to your session.
      parameters:
        - { name: publisherId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Dashboard totals, endpoints, transactions and settlements }
        '403': { description: Not your dashboard }
  /api/dashboard/{publisherId}/calls:
    get:
      tags: [Dashboard]
      summary: Every call with the transaction that paid for it
      description: |
        Reconcile this against your own upstream access logs. A request your server saw
        that is not listed here, or a row with no transaction, is a discrepancy — the
        gateway settles before proxying, so an unpaid call cannot occur by design.
      parameters:
        - { name: publisherId, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        '200':
          description: Verifiable call log
          content:
            application/json:
              schema:
                type: object
                properties:
                  calls:
                    type: array
                    items: { $ref: '#/components/schemas/CallRecord' }
                  unpaidCalls:
                    type: integer
                    description: Should always be 0. Anything else is worth reporting.
  /api/dashboard/{publisherId}/settle:
    post:
      tags: [Dashboard]
      summary: Pay your balance to your own wallet
      parameters:
        - { name: publisherId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Paid
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [PAID] }
                  txId: { type: string }
                  explorerUrl: { type: string }
        '400': { description: Nothing due, or below the minimum payout }
        '409': { description: No payout address set }

  # ------------------------------------------------------------- Gateway ----
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: x402_session
      description: Set by sign-in. Used by browsers.
    bearerToken:
      type: http
      scheme: bearer
      description: The same session token, for CLI and server-to-server clients.
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
...
  schemas:
...
    CallRecord:
      type: object
      properties:
        created_at: { type: string, format: date-time }
        method: { type: string }
        path: { type: string }
        status_code: { type: integer }
        latency_ms: { type: integer }
        publisher_revenue: { type: number }
        txId: { type: string, nullable: true }
        explorerUrl: { type: string, nullable: true }